metacognition-sdk/pyproject.toml
[project]
name = "metacognition-sdk"
version = "0.1.0"
description = ""
authors = ["Plastic Labs <hello@plasticlabs.ai>"]
readme = "README.md"

[tool.poetry]
name = "metacognition-sdk"
version = "0.1.0"
description = ""
authors = ["Plastic Labs <hello@plasticlabs.ai>"]
readme = "README.md"
packages = [{include = "metacognition_sdk"}]

[tool.poetry.dependencies]
python = "^3.8"
tqdm = "^4.66.1"
rich = "^13.7.0"
openai = "^1.9.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
